Moving test stubs to their own file
